organizing metric and pre-processing functions
